[MOV] evaluation: move file in its own folder
This commit prepares the array formula evaluation (next commit)
by moving the evaluation plugin file to its own folder
