+++ src/plugins/ui_core_views/cell_evaluation/index.ts
@@ -1,0 +1,1 @@
+export * from "./evaluation_plugin";
