+++ src/plugins/ui_core_views/index.ts
@@ -1,5 +1,5 @@
+export * from "./cell_evaluation";
 export * from "./custom_colors";
-export * from "./evaluation";
 export * from "./evaluation_chart";
 export * from "./evaluation_conditional_format";
 export * from "./filter_evaluation";
